refactor(chat): implement new error handling utilities and streamline interrupt request processing in NewChatPage for improved performance and maintainability
